Testing app comp. with Jest
