+++ src/pages/main/index.jsx
@@ -1,14 +1,33 @@
-import React from "react";
+import React, { Component } from "react";
 import Calendar from "../../components/Calendar/index";
 import Notes from "../../components/Notes/index";
 
-const Main = () => {
-  return (
-    <div className="container">
-      <Calendar></Calendar>
-      <Notes></Notes>
-    </div>
-  );
-};
+import { connect } from "react-redux";
+import { bindActionCreators } from "redux";
 
-export default Main;
+import { Creators as CalendarActions } from "../../store/ducks/calendar";
+
+class Main extends Component {
+  componentDidMount() {
+    const { requestLoadCalendar, searchDate } = this.props;
+    requestLoadCalendar(searchDate);
+  }
+
+  render() {
+    return (
+      <div className="container">
+        <Calendar></Calendar>
+        <Notes></Notes>
+      </div>
+    );
+  }
+}
+
+const mapStateToProps = (state) => ({
+  searchDate: state.calendar.searchDate,
+});
+
+const mapDispatchToProps = (dispatch) =>
+  bindActionCreators(CalendarActions, dispatch);
+
+export default connect(mapStateToProps, mapDispatchToProps)(Main);
